backups/object-backup/requirements.txt
boto3==1.20.23
PyYAML==5.4.1
influxdb==5.3.1
